requirements.txt
fastapi
uvicorn
openai==0.28
python-dotenv
requests
duckduckgo-search
